% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funs.R
\name{desc_quant}
\alias{desc_quant}
\alias{desc_freq}
\title{Descriptive statistics and frequency tables}
\usage{
desc_quant(.data, ..., .by = NULL, .fns = .desc_quant_fns)

desc_freq(.data, ..., .by = NULL)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble)}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}

\item{.by}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of
columns to group by for just this operation.}

\item{.fns}{Named list of functions to apply to the columns specified in
\code{...}. The names of the list will be the column names of the resulting
data frame.}
}
\value{
An object \emph{usually} of the same type as \code{.data}.

Each column is the result of a function defined in \code{.fns}. Each
row is a grouping variable defined in \code{.by}, if any.
}
\description{
Convenient functions for describing data.
\itemize{
\item \code{desc_quant} runs quantitative descriptives of quantitative data.
\item \code{desc_freq} runs frequency statistics for categorical data.
}
}
\examples{
# no examples available
}
