% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Title}
\usage{
describe(
  .data,
  ...,
  .quant_funs = list(M = ~mean(., na.rm = TRUE), SD = ~sd(., na.rm = TRUE), N =
    ~sum(!is.na(.)))
)
}
\arguments{
\item{.quant_funs}{}
}
\value{

}
\description{
Title
}
\examples{

}
