% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{latest}
\alias{latest}
\title{Get path of latest matched file}
\usage{
latest(pattern, path = ".", ...)
}
\arguments{
\item{pattern}{Regular expression pattern}

\item{path}{A character vector of full path names; the default corresponds
to the working directory, \code{\link[=getwd]{getwd()}}. Tilde expansion (see \link{path.expand}) is
performed. Missing values will be ignored. Elements with a marked encoding
will be converted to the native encoding (and if that fails, considered
non-existent).}

\item{...}{Additional parameters passed to \code{dir}.}
}
\value{
A character vector of length 1 that contains the most recent file
that matches \code{pattern}.

An error is thrown if the \code{pattern} cannot be matched. A confirmation
message is printed if the \code{pattern} is found.
}
\description{
This function searches for the latest file that matches a search pattern in
a given directory. This is particular useful for exported datasets that
include ever changing download times in their filename.
}
