% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spss_ops.R
\name{make_spss_ops}
\alias{make_spss_ops}
\title{Make SPSS operations for a mutation}
\usage{
make_spss_ops(.table, na_values = -99)
}
\arguments{
\item{.table}{A list or data frame with three elements:
\enumerate{
\item Character vector of variable names
\item Character vector of variable labels
\item Character vector of value labels, each element of the form \code{label1 = value1, label2 = value2, ...}
}}

\item{na_values}{Values to be marked as \code{NA} in SPSS. Applied to all variables given by \code{.table}.}
}
\value{
A list of quosures to be spliced into a \code{mutate} function.
}
\description{
Making SPSS operations for a mutation
}
